move crypto lib to client directory
