+++ client/cmdhfmfp.c
@@ -24,7 +24,7 @@
 #include "mifare.h"
 #include "mifare4.h"
 #include "cliparser/cliparser.h"
-#include "polarssl/libpcrypto.h"
+#include "crypto/libpcrypto.h"
 
 static const uint8_t DefaultKey[16] = {0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff};
